Update ModHashFromName to always lowercase first

+++ src/dbg/murmurhash.h
@@ -36,16 +36,16 @@ void MurmurHash3_x64_128(const void* key, int len, uint32_t seed, void* out);
 
 #ifdef _WIN64
 static inline
-unsigned long long murmurhash(const void* data, int len)
+unsigned long long murmurhash(const void* data, size_t len)
 {
     unsigned long long hash[2];
-    MurmurHash3_x64_128(data, len, 0x1337, hash);
+    MurmurHash3_x64_128(data, (int)len, 0x1337, hash);
 #else //x86
 static inline
-unsigned long murmurhash(const void* data, int len)
+unsigned long murmurhash(const void* data, size_t len)
 {
     unsigned int hash[1];
-    MurmurHash3_x86_32(data, len, 0x1337, hash);
+    MurmurHash3_x86_32(data, (int)len, 0x1337, hash);
 #endif //_WIN64
     return hash[0];
 }
